requirements_macos_arm.txt
# This requirements file specifies the dependencies needed to run the project on
# macOS devices with Apple Silicon (M1/M2/M3). It includes optimized versions of
# TensorFlow for ARM architecture, ensuring improved performance on Apple GPUs
# via tensorflow-macos and tensorflow-metal.  
#  
# The listed libraries cover data processing, visualization, machine learning,  
# hyperparameter optimization, and progress monitoring.  
#  
# Create your environment using:  
#   python3 -m venv .venv  
#   source .venv/bin/activate  
#   pip install -r requirements_macos_arm.txt

numpy>=1.26
pandas>=2.2
matplotlib>=3.9
seaborn>=0.13
scikit-learn>=1.4
tqdm>=4.66
optuna>=3.6.1

# TensorFlow optimized for Apple Silicon
tensorflow-macos>=2.16
tensorflow-metal>=1.1